# api/openapi.yaml
openapi: 3.1.0
info:
  title: Arca
  version: "1"
  description: "Durable storage for people, agents, and sandboxes: files with versions, trash and stars, uploads in parts, shares with a permission ladder and public links, workspaces with a writer lease, and an event log."
paths:
...
  /v1/events:
    get:
      operationId: getV1Events
      summary: List events
      description: One page of a space's log, oldest first. Asks the authorizer for event.read before it acts.
      security:
      - bearer: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: {}
        "401":
          $ref: "#/components/responses/unauthenticated"
        "403":
          $ref: "#/components/responses/forbidden"
        "404":
          $ref: "#/components/responses/not_found"
        "429":
          $ref: "#/components/responses/rate_limited"
        "500":
          $ref: "#/components/responses/internal"
        "503":
          description: "authorizer_unavailable: The permission service is unavailable; retry shortly."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /v1/shares:
    get:
      operationId: getV1Shares
      summary: List grants
      description: The grants on a space. Asks the authorizer for share.list before it acts.
      security:
      - bearer: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: {}
        "401":
          $ref: "#/components/responses/unauthenticated"
        "403":
          $ref: "#/components/responses/forbidden"
        "404":
          $ref: "#/components/responses/not_found"
        "429":
          $ref: "#/components/responses/rate_limited"
        "500":
          $ref: "#/components/responses/internal"
        "503":
          description: "authorizer_unavailable: The permission service is unavailable; retry shortly."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      operationId: postV1Shares
      summary: Create grant
      description: Grant a subject a permission on a subtree of a space. Asks the authorizer for share.create before it acts.
      security:
      - bearer: []
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: {}
        "401":
          $ref: "#/components/responses/unauthenticated"
        "403":
          $ref: "#/components/responses/forbidden"
        "404":
          $ref: "#/components/responses/not_found"
        "429":
          $ref: "#/components/responses/rate_limited"
        "500":
          $ref: "#/components/responses/internal"
        "503":
          description: "authorizer_unavailable: The permission service is unavailable; retry shortly."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /v1/shares/links:
    get:
      operationId: getV1SharesLinks
      summary: List links
      description: The token grants on a space. Asks the authorizer for link.read before it acts.
      security:
      - bearer: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: {}
        "401":
          $ref: "#/components/responses/unauthenticated"
        "403":
          $ref: "#/components/responses/forbidden"
        "404":
          $ref: "#/components/responses/not_found"
        "429":
          $ref: "#/components/responses/rate_limited"
        "500":
          $ref: "#/components/responses/internal"
        "503":
          description: "authorizer_unavailable: The permission service is unavailable; retry shortly."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      operationId: postV1SharesLinks
      summary: Create link
      description: Mint a token grant on a subtree; the token is answered once. Asks the authorizer for link.create before it acts.
      security:
      - bearer: []
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: {}
        "401":
          $ref: "#/components/responses/unauthenticated"
        "403":
          $ref: "#/components/responses/forbidden"
        "404":
          $ref: "#/components/responses/not_found"
        "429":
          $ref: "#/components/responses/rate_limited"
        "500":
          $ref: "#/components/responses/internal"
        "503":
          description: "authorizer_unavailable: The permission service is unavailable; retry shortly."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /v1/shares/links/{id}:
    delete:
      operationId: deleteV1SharesLinksById
      summary: Revoke link
      description: Revoke a token grant. Asks the authorizer for link.revoke before it acts.
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      security:
      - bearer: []
      responses:
        "204":
          description: No Content
          content:
            application/json:
              schema: {}
        "401":
          $ref: "#/components/responses/unauthenticated"
        "403":
          $ref: "#/components/responses/forbidden"
        "404":
          $ref: "#/components/responses/not_found"
        "429":
          $ref: "#/components/responses/rate_limited"
        "500":
          $ref: "#/components/responses/internal"
        "503":
          description: "authorizer_unavailable: The permission service is unavailable; retry shortly."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /v1/shares/links/{token}:
    get:
      operationId: getV1SharesLinksByToken
      summary: List linked files
      description: "A listing of the subtree a link token names. Carries no bearer token: the token in the URL is the whole of the authorization."
      parameters:
      - name: token
        in: path
        required: true
        schema:
          type: string
      security: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: {}
        "429":
          $ref: "#/components/responses/rate_limited"
        "500":
          $ref: "#/components/responses/internal"
  /v1/shares/links/{token}/files/{path}:
    get:
      operationId: getV1SharesLinksByTokenFilesByPath
      summary: Read linked file
      description: "One object under the subtree a link token names. Carries no bearer token: the token in the URL is the whole of the authorization."
      parameters:
      - name: token
        in: path
        required: true
        schema:
          type: string
      - name: path
        in: path
        required: true
        description: The rest of the path, slashes included.
        schema:
          type: string
      security: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: {}
        "429":
          $ref: "#/components/responses/rate_limited"
        "500":
          $ref: "#/components/responses/internal"
  /v1/shares/links/{token}/meta:
    get:
      operationId: getV1SharesLinksByTokenMeta
      summary: Get link metadata
      description: "What a link token names, before anything is fetched. Carries no bearer token: the token in the URL is the whole of the authorization."
      parameters:
      - name: token
        in: path
        required: true
        schema:
          type: string
      security: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: {}
        "429":
          $ref: "#/components/responses/rate_limited"
        "500":
          $ref: "#/components/responses/internal"
  /v1/shares/with-me:
    get:
      operationId: getV1SharesWithMe
      summary: List received grants
      description: The grants whose grantee is the caller. Asks the authorizer for share.list before it acts.
      security:
      - bearer: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: {}
        "401":
          $ref: "#/components/responses/unauthenticated"
        "403":
          $ref: "#/components/responses/forbidden"
        "404":
          $ref: "#/components/responses/not_found"
        "429":
          $ref: "#/components/responses/rate_limited"
        "500":
          $ref: "#/components/responses/internal"
        "503":
          description: "authorizer_unavailable: The permission service is unavailable; retry shortly."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /v1/shares/{id}:
    delete:
      operationId: deleteV1SharesById
      summary: Revoke grant
      description: Revoke a grant. Asks the authorizer for share.revoke before it acts.
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      security:
      - bearer: []
      responses:
        "204":
          description: No Content
          content:
            application/json:
              schema: {}
        "401":
          $ref: "#/components/responses/unauthenticated"
        "403":
          $ref: "#/components/responses/forbidden"
        "404":
          $ref: "#/components/responses/not_found"
        "429":
          $ref: "#/components/responses/rate_limited"
        "500":
          $ref: "#/components/responses/internal"
        "503":
          description: "authorizer_unavailable: The permission service is unavailable; retry shortly."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    get:
      operationId: getV1SharesById
      summary: Get grant
      description: One grant. Asks the authorizer for share.read before it acts.
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      security:
      - bearer: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: {}
        "401":
          $ref: "#/components/responses/unauthenticated"
        "403":
          $ref: "#/components/responses/forbidden"
        "404":
          $ref: "#/components/responses/not_found"
        "429":
          $ref: "#/components/responses/rate_limited"
        "500":
          $ref: "#/components/responses/internal"
        "503":
          description: "authorizer_unavailable: The permission service is unavailable; retry shortly."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: A token from an issuer this installation lists, carrying the audience it verifies.
  schemas:
    Error:
      type: object
      description: The one error shape of this API. A caller branches on error.code and shows error.message.
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              description: The stable identifier of the row of the error table.
            details:
              type: object
              description: "What varies: always request_id, the developer sentence in detail, and the field paths at fault in fields."
              properties:
                detail:
                  type: string
                fields:
                  type: array
                  items:
                    type: string
                request_id:
                  type: string
            message:
              type: string
              description: The one sentence of that row, fixed and shown to a person.
          required:
          - code
          - message
      required:
      - error
...
  responses:
...
    forbidden:
      description: "403 forbidden: You do not have permission to do this."
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    internal:
      description: "500 internal: Something went wrong on this server."
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    not_found:
      description: "404 not_found: There is no such object."
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    rate_limited:
      description: "429 rate_limited: Too many requests; wait and retry."
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    unauthenticated:
      description: "401 unauthenticated: Sign in and send a valid token."
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...